Book Maker/requirements.txt
openai>=1.0.0
pandas>=1.5.0
python-docx>=0.8.11
pydub>=0.25.1
scikit-learn>=1.0.0
sentence-transformers>=2.0.0
numpy>=1.21.0
openpyxl>=3.0.0

